+++ src/components/article/Article.tsx
@@ -1,5 +1,3 @@
-import clsx from 'clsx';
-
 // Сообщаем вебпаку, что этот файл использует это изображение.
 import plane from 'src/images/plane.png';
 import { Text } from 'components/text';
@@ -8,7 +6,7 @@ import styles from './Article.module.scss';
 
 export const Article = () => {
 	return (
-		<article className={clsx(styles.article)}>
+		<article className={styles.article}>
 			<Text as='h1' size={45} weight={800} uppercase dynamicLite>
 				Портрет Западной Швейцарии
 			</Text>
